requirements.txt
numpy==1.23.1
PyYAML==6.0
requests==2.25.1
scipy==1.8.1
tqdm==4.42.1
